Use 64-bit delta for Datetime +/- operators
This allows larger deltas, as `int` is only 32 bits. Even on platforms
where time_t is 32 bits, using `int64_t` is not worse than using `int`.

+++ src/Datetime.cpp
@@ -3788,26 +3788,26 @@ bool Datetime::sameYear (const Datetime& rhs) const
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-Datetime Datetime::operator+ (const int delta)
+Datetime Datetime::operator+ (const int64_t delta)
 {
   return { _date + delta };
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-Datetime Datetime::operator- (const int delta)
+Datetime Datetime::operator- (const int64_t delta)
 {
   return { _date - delta };
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-Datetime& Datetime::operator+= (const int delta)
+Datetime& Datetime::operator+= (const int64_t delta)
 {
   _date += (time_t) delta;
   return *this;
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-Datetime& Datetime::operator-= (const int delta)
+Datetime& Datetime::operator-= (const int64_t delta)
 {
   _date -= (time_t) delta;
   return *this;
